bsc: Work on MGCP and resource assignment
This will allocate a circuit on creation of a CC Setup message
and it will tear it down when the connection is going way. It
does not support MDCX or such as this will require queuing. It
also needs to be moved to a class of its own and be hooked with
a remote party...
The most confusing part is how this will morph for call holding
and such as there is only one assignment but we would have multiple
users for the audio.