bsc: Send the assignment command first, then do MGCP
The codec selection is part of the assignment command, this is
means we can only do MGCP handling after the BSC has selected
the audio codecs and assigned the circuits. So start with the
GSM0808 assigment command, this means that we might have a MGCP
Endpoint that is reserved but doesn't have a CallID, we handle
this in the releaseAudio and free the endpoint.