cm: Deal with CM Service Requests to a certain degree.
The CM Service Request accepts the service and then we need to
wait for the real service. It could check if the requested service
is matching with the 'inital' message but this is not done yet.

The biggest issue is in the hand-over from CMServiceRequest to the
real request. Right now a second SMS submit would break things
up.