sip: Assign an MSC identity to a call
For some usecases we have a list of identities and when selecting
a route an unused identity/line needs to be used. Keep track of
which lines/identities we are using.

The code might release a line too early. There should probably
be another call in Osmo.SIPCall to inform the code when the call
has "terminated" in one way or another (busy, cancel, failure,
remote hangup, local hangup, redirected, etc). Right now it could
still happen than an identity remains used.